refactor: Migrate auth-generator to typescript (LPS-119618)
